requirements.txt
python-dotenv
PyAudio
Pillow
mss
google-generativeai